+++ src/map.rs
@@ -89,6 +89,12 @@ impl<T, C> SkipMap<T, C> {
     self.len() == 0
   }
 
+  /// Returns the maximum version of all entries in the map.
+  #[inline]
+  pub fn max_version(&self) -> u64 {
+    self.arena.max_version.load(Ordering::Acquire)
+  }
+
   /// Flushes outstanding memory map modifications to disk.
   ///
   /// When this method returns with a non-error result,
@@ -400,6 +406,7 @@ impl<T: Trailer, C: Comparator> SkipMap<T, C> {
       return Err(Error::Readonly);
     }
 
+    let version = trailer.version();
     self
       .insert_in::<Infallible>(
         trailer,
@@ -411,6 +418,30 @@ impl<T: Trailer, C: Comparator> SkipMap<T, C> {
         },
         &mut Inserter::default(),
       )
+      .map(|ent| {
+        // Update the max version.
+        if ent.is_none() {
+          let mut current = self.arena.max_version.load(Ordering::Acquire);
+          loop {
+            match self.arena.max_version.compare_exchange_weak(
+              current,
+              version,
+              Ordering::SeqCst,
+              Ordering::Acquire,
+            ) {
+              Ok(_) => break,
+              Err(v) => {
+                if v < version {
+                  current = v;
+                } else {
+                  break;
+                }
+              }
+            }
+          }
+        }
+        ent
+      })
       .map_err(|e| e.expect_right("must be map::Error"))
   }