+++ src/map/tests.rs
@@ -1643,20 +1643,22 @@ fn test_reopen_mmap2() {
   {
     let l = SkipMap::mmap_mut_with_comparator(&p, ARENA_SIZE, true, Ascend).unwrap();
     for i in 0..1000 {
-      l.insert(0, &key(i), &new_value(i)).unwrap();
+      l.insert(i as u64, &key(i), &new_value(i)).unwrap();
     }
     l.flush_async().unwrap();
+    assert_eq!(l.max_version(), 999);
   }
 
   let l = SkipMap::<u64, Ascend>::mmap_with_comparator(&p, false, Ascend).unwrap();
   assert_eq!(1000, l.len());
   for i in 0..1000 {
     let k = key(i);
-    let ent = l.get(0, &k).unwrap();
+    let ent = l.get(i as u64, &k).unwrap();
     assert_eq!(new_value(i), ent.value());
-    assert_eq!(ent.trailer().version(), 0);
+    assert_eq!(ent.trailer().version(), i as u64);
     assert_eq!(ent.key(), k);
   }
+  assert_eq!(l.max_version(), 999);
 }
 
 struct Person {